Deduplicate magnet helpers and fix spread order in providers
Extract shared extractInfoHash, base32ToHex, parseSize into
scraper/lib/magnetHelper.js. Remove duplicated local copies from
therarbg, subsplease, animetosho, and nekobt providers.

Fix object spread order so parseTitle() is spread first and
provider-specific fields (quality, languages, source) override
its defaults instead of being overwritten.

Align import spacing in providers/index.js.
